requirements.txt
flask
flask-cors
ultralytics
opencv-python-headless
Pillow
numpy
